+++ api/src/metrics.rs
@@ -627,7 +627,6 @@ macro_rules! define_sync_instrument {
             attribute_count: usize,
         ) -> OtelStatus {
             guard_status(|| {
-                clear_last_error();
                 let instrument = match unsafe { checked_ref(instrument) } {
                     Some(instrument) => instrument,
                     None => return OtelStatus::InvalidArgument,
@@ -635,6 +634,7 @@ macro_rules! define_sync_instrument {
                 if instrument.vtable.is_null() {
                     return OtelStatus::Ok;
                 }
+                clear_last_error();
                 unsafe {
                     ((*instrument.vtable).$vtable_record)(
                         instrument.ctx,
@@ -887,7 +887,6 @@ macro_rules! define_bound_instrument {
             value: $value,
         ) -> OtelStatus {
             guard_status(|| {
-                clear_last_error();
                 let instrument = match unsafe { checked_ref(instrument) } {
                     Some(instrument) => instrument,
                     None => return OtelStatus::InvalidArgument,
@@ -895,6 +894,7 @@ macro_rules! define_bound_instrument {
                 if instrument.vtable.is_null() {
                     return OtelStatus::Ok;
                 }
+                clear_last_error();
                 unsafe { ((*instrument.vtable).$vtable_record)(instrument.ctx, value) }
             })
         }